+++ libvast/src/schema.cpp
@@ -255,35 +255,48 @@ caf::expected<schema> load_schema(const path& schema_file) {
   return to<schema>(*str);
 }
 
+caf::error load_symbols(const path& schema_file, symbol_table& local) {
+  if (schema_file.empty())
+    return caf::make_error(ec::filesystem_error, "empty path");
+  auto str = load_contents(schema_file);
+  if (!str)
+    return str.error();
+  auto p = symbol_table_parser{};
+  if (!p(*str, local))
+    return caf::make_error(ec::parse_error, "failed to load symbols from",
+                           schema_file);
+  return caf::none;
+}
+
 caf::expected<schema>
 load_schema(const detail::stable_set<path>& schema_dirs, size_t max_recursion) {
   if (max_recursion == 0)
     return ec::recursion_limit_reached;
   vast::schema types;
+  symbol_table global_symbols;
   for (const auto& dir : schema_dirs) {
     VAST_VERBOSE("loading schemas from {}", dir);
     if (!exists(dir)) {
       VAST_DEBUG("{} skips non-existing directory: {}", __func__, dir);
       continue;
     }
-    vast::schema directory_schema;
     auto filter
       = [](const path& f) { return detail::ends_with(f.str(), ".schema"); };
     auto schema_files = filter_dir(dir, std::move(filter), max_recursion);
-    for (auto f : schema_files) {
+    symbol_table local_symbols;
+    for (const auto& f : schema_files) {
       VAST_DEBUG("loading schema {}", f);
-      auto schema = load_schema(f);
-      if (!schema) {
-        VAST_ERROR("{} {} {}", __func__, render(schema.error()), f);
-        continue;
-      }
-      if (auto merged = schema::merge(directory_schema, *schema))
-        directory_schema = std::move(*merged);
-      else
-        return caf::make_error(ec::format_error, merged.error().context(),
-                               "in schema file", f);
+      if (auto err = load_symbols(f, local_symbols))
+        return err;
     }
-    types = schema::combine(types, directory_schema);
+    auto r = symbol_resolver{global_symbols, local_symbols};
+    auto directory_schema = r.resolve();
+    if (!directory_schema)
+      return caf::make_error(ec::format_error, "failed to resolve types in",
+                             dir);
+    local_symbols.merge(std::move(global_symbols));
+    global_symbols = std::move(local_symbols);
+    types = schema::combine(types, *directory_schema);
   }
   return types;
 }